+++ packages/api/src/index.ts
@@ -1,26 +1,35 @@
 export * from './api.gen'
 
+import { ArcadeumAPI as BaseArcadeumAPI } from './api.gen'
 import fetch from 'cross-fetch'
 
-import { ArcadeumAPI as BaseArcadeumAPI } from './api.gen'
-
 export class ArcadeumAPIClient extends BaseArcadeumAPI {
-  constructor(hostname: string, public jwtAuth?: string) {
+  jwtAuth: string | undefined = undefined
+
+  constructor(hostname: string) {
     super(hostname, fetch)
     this.fetch = this._fetch
   }
 
   _fetch = (input: RequestInfo, init?: RequestInit): Promise<Response> => {
-    // automatically include jwt auth header to requests
-    // if its been set on the api client
-    const headers: { [key: string]: any } = {}
-    if (this.jwtAuth && this.jwtAuth.length > 0) {
-      headers['Authorization'] = `BEARER ${this.jwtAuth}`
-    }
+    return new Promise<Response>((resolve, reject) => {
+      // automatically include jwt auth header to requests
+      // if its been set on the api client
+      const headers: { [key: string]: any } = {}
+      if (this.jwtAuth && this.jwtAuth.length > 0) {
+        headers['Authorization'] = `BEARER ${this.jwtAuth}`
+      }
 
-    // before the request is made
-    init!.headers = { ...init!.headers, ...headers }
-
-    return fetch(input, init)
-  }
+      // before the request is made
+      init!.headers = { ...init!.headers, ...headers }
+  
+      fetch(input, init).then(resp => {
+        // after the request has been made..
+        resolve(resp)
+      }).catch(err => {
+        // request error
+        reject(err)
+      })
+    })
+  } 
 }