+++ packages/auth/tests/session.spec.ts
@@ -30,48 +30,6 @@ type EthereumInstance = {
   signer?: AbstractSigner
 }
 
-class CountingSigner extends AbstractSigner {
-  private _signingRequests: number = 0
-
-  constructor(private readonly signer: AbstractSigner) {
-    super()
-  }
-
-  get signingRequests(): number {
-    return this._signingRequests
-  }
-
-  getAddress(): Promise<string> {
-    return this.signer.getAddress()
-  }
-
-  signMessage(message: ethers.Bytes | string): Promise<string> {
-    this._signingRequests++
-    return this.signer.signMessage(message)
-  }
-
-  signTransaction(transaction: ethers.utils.Deferrable<ethers.providers.TransactionRequest>): Promise<string> {
-    this._signingRequests++
-    return this.signer.signTransaction(transaction)
-  }
-
-  connect(provider: ethers.providers.Provider): ethers.Signer {
-    return this.signer.connect(provider)
-  }
-}
-
-async function initialAuthToFinish(session: Session): Promise<void> {
-  await Promise.allSettled(session._initialAuthRequests)
-}
-
-function hasTokenForURL(session: Session, url: string): boolean {
-  return session._jwts.has(url)
-}
-
-function getTokenForURL(session: Session, url: string): Promise<string | undefined> {
-  return session._jwts.get(url)?.token
-}
-
 describe('Wallet integration', function () {
   const ethnode: EthereumInstance = {}
 
@@ -162,7 +120,7 @@ describe('Wallet integration', function () {
   it("Should dump and load a session", async () => {
     const referenceSigner = ethers.Wallet.createRandom()
 
-    let session = await Session.open({
+    const dump = (await Session.open({
       context: context,
       networks: networks,
       referenceSigner: referenceSigner.address,
@@ -171,11 +129,9 @@ describe('Wallet integration', function () {
       metadata: {
         name: "Test"
       }
-    })
+    })).dump()
 
-    const dump = await session.dump()
-
-    session = Session.load({
+    const session = Session.load({
       dump: dump,
       signers: [referenceSigner],
       networks: networks
@@ -455,10 +411,10 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       await session.auth(networksWithApi[0])
       expect(totalCount).to.equal(1)
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
       expect(proofAddress).to.equal(session.account.address)
     })
 
@@ -475,10 +431,10 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       await session.auth(ethnode.chainId)
       expect(totalCount).to.equal(1)
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
       expect(proofAddress).to.equal(session.account.address)
     })
 
@@ -495,7 +451,7 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const auth = session.auth({
         name: 'mainnet',
         chainId: 1
@@ -517,7 +473,7 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const auth = session.auth(1)
 
       await expect(auth).to.be.rejected
@@ -536,9 +492,9 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const newSigner = ethers.Wallet.createRandom()
-
+  
       const session = await Session.open({
         context: context,
         networks: networks,
@@ -551,10 +507,10 @@ describe('Wallet integration', function () {
       })
 
       await session.auth(networksWithApi[0])
-      await initialAuthToFinish(session)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
       expect(totalCount).to.equal(1)
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
       expect(proofAddress).to.equal(session.account.address)
     })
 
@@ -572,13 +528,13 @@ describe('Wallet integration', function () {
         }
       })
 
-      expect(session._initialAuthRequests.length).to.equal(networksWithApi.length)
-      await initialAuthToFinish(session)
+      expect(session.authPromises.length).to.equal(1)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
       expect(totalCount).to.equal(1)
       expect(recoverCount[session.account.address]).to.equal(1)
 
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
     })
 
     it("Should get JWT during session opening", async () => {
@@ -586,7 +542,7 @@ describe('Wallet integration', function () {
 
       const referenceSigner = ethers.Wallet.createRandom()
 
-      let session = await Session.open({
+      await Session.open({
         context: context,
         networks: networks,
         referenceSigner: referenceSigner.address,
@@ -596,12 +552,10 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
-      await initialAuthToFinish(session)
-
+  
       const newSigner = ethers.Wallet.createRandom()
-
-      session = await Session.open({
+  
+      const session = await Session.open({
         context: context,
         networks: networksWithApi,
         referenceSigner: referenceSigner.address,
@@ -611,16 +565,15 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
+  
+      expect(session.authPromises.length).to.equal(2)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
-      expect(session._initialAuthRequests.length).to.equal(networksWithApi.length)
-      await initialAuthToFinish(session)
-
-      expect(totalCount).to.equal(1)
-      // TODO: can't reproduce
-      // expect(recoverCount["error"]).to.equal(1)
+      expect(totalCount).to.equal(2)
+      expect(recoverCount["error"]).to.equal(1)
       expect(recoverCount[session.account.address]).to.equal(1)
 
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
     })
 
     it("Should get API with lazy JWT during first session creation", async () => {
@@ -642,7 +595,7 @@ describe('Wallet integration', function () {
       expect(totalCount).to.equal(1)
       expect(recoverCount[session.account.address]).to.equal(1)
 
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
 
       server.post('/rpc/ArcadeumAPI/FriendList').thenCallback(async (request) => {
         const hasToken = request.headers['authorization'].includes(fakeJwt)
@@ -666,9 +619,9 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const newSigner = ethers.Wallet.createRandom()
-
+  
       const session = await Session.open({
         context: context,
         networks: networks,
@@ -685,7 +638,7 @@ describe('Wallet integration', function () {
       expect(totalCount).to.equal(1)
       expect(recoverCount[session.account.address]).to.equal(1)
 
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
 
       server.post('/rpc/ArcadeumAPI/FriendList').thenCallback(async (request) => {
         const hasToken = request.headers['authorization'].includes(fakeJwt)
@@ -712,7 +665,7 @@ describe('Wallet integration', function () {
       let calledCallback = 0
       session.onAuth(() => calledCallback++)
 
-      await initialAuthToFinish(session)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
       expect(calledCallback).to.equal(1)
     })
@@ -732,9 +685,9 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const newSigner = ethers.Wallet.createRandom()
-
+  
       const session = await Session.open({
         context: context,
         networks: networksWithApi,
@@ -749,7 +702,7 @@ describe('Wallet integration', function () {
       let calledCallback = 0
       session.onAuth(() => calledCallback++)
 
-      await initialAuthToFinish(session)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
       expect(calledCallback).to.equal(1)
     })
@@ -773,7 +726,7 @@ describe('Wallet integration', function () {
       const auth = session.auth(networksWithApi[0])
       await expect(auth).to.be.rejected
       expect(totalCount).to.equal(5)
-      expect(hasTokenForURL(session, sequenceApiUrl)).to.be.false
+      expect(session.jwts[sequenceApiUrl]).to.be.undefined
     })
 
     it("Should get API with JWT already present", async () => {
@@ -790,9 +743,9 @@ describe('Wallet integration', function () {
           name: "Test"
         }
       })
-
+  
       const newSigner = ethers.Wallet.createRandom()
-
+  
       const session = await Session.open({
         context: context,
         networks: networksWithApi,
@@ -804,19 +757,15 @@ describe('Wallet integration', function () {
         }
       })
 
-      await initialAuthToFinish(session)
+      await Promise.all(session.authPromises.map((p) => p.promise))
 
-      const totalCountBefore = totalCount
-
-      // This should use the already existing JWT
       const api = await session.getAPI(networksWithApi[0])
 
-      expect(totalCount).to.equal(totalCountBefore)
-      // TODO: can't reproduce
-      // expect(recoverCount["error"]).to.equal(1)
+      expect(totalCount).to.equal(2)
+      expect(recoverCount["error"]).to.equal(1)
       expect(recoverCount[session.account.address]).to.equal(1)
 
-      expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
+      expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
 
       server.post('/rpc/ArcadeumAPI/FriendList').thenCallback(async (request) => {
         const hasToken = request.headers['authorization'].includes(fakeJwt)
@@ -829,8 +778,6 @@ describe('Wallet integration', function () {
     it("Should fail to get API with false tryAuth and no JWT", async () => {
       const referenceSigner = ethers.Wallet.createRandom()
 
-      alwaysFail = true
-
       const session = await Session.open({
         context: context,
         networks: networks,
@@ -842,16 +789,12 @@ describe('Wallet integration', function () {
         }
       })
 
-      await initialAuthToFinish(session)
-
-      alwaysFail = false
-
       const apiPromise = session.getAPI(networksWithApi[0], false)
 
       await expect(apiPromise).to.be.rejected
 
       expect(totalCount).to.equal(0)
-      expect(hasTokenForURL(session, sequenceApiUrl)).to.be.false
+      expect(session.jwts[sequenceApiUrl]).to.be.undefined
     })
 
     it("Should fail to get API without api url", async () => {
@@ -873,7 +816,7 @@ describe('Wallet integration', function () {
       await expect(apiPromise).to.be.rejected
 
       expect(totalCount).to.equal(0)
-      expect(hasTokenForURL(session, sequenceApiUrl)).to.be.false
+      expect(session.jwts[sequenceApiUrl]).to.be.undefined
     })
 
     it("Should fail to get JWT with no api configured", async () => {
@@ -895,109 +838,7 @@ describe('Wallet integration', function () {
       await expect(authPromise).to.be.rejected
 
       expect(totalCount).to.equal(0)
-      expect(hasTokenForURL(session, sequenceApiUrl)).to.be.false
-    })
-
-    it('Should reuse outstanding JWT requests', async () => {
-      const referenceSigner = new CountingSigner(ethers.Wallet.createRandom())
-
-      alwaysFail = true
-
-      const session = await Session.open({
-        context,
-        networks: networksWithApi,
-        referenceSigner: await referenceSigner.getAddress(),
-        signers: [{ signer: referenceSigner, weight: 1 }],
-        thershold: 1,
-        metadata: {
-          name: 'Test'
-        }
-      })
-
-      // 2 signatures are made to publish signers
-      expect(referenceSigner.signingRequests).to.equal(2)
-
-      const signingRequestsBefore = referenceSigner.signingRequests
-
-      await initialAuthToFinish(session)
-
-      alwaysFail = false
-      totalCount = 0
-
-      // Create a bunch of API clients concurrently
-      const requests = []
-      while (requests.length < 10) {
-        requests.push(session.getAPI(networksWithApi[0]))
-      }
-      await expect(Promise.all(requests)).to.be.fulfilled
-
-      expect(totalCount).to.equal(1)
-      expect(referenceSigner.signingRequests).to.equal(signingRequestsBefore + 1)
-    })
-
-    it('Should reuse existing proof signatures', async () => {
-      const referenceSigner = new CountingSigner(ethers.Wallet.createRandom())
-
-      alwaysFail = true
-
-      const session = await Session.open({
-        context,
-        networks: networksWithApi,
-        referenceSigner: await referenceSigner.getAddress(),
-        signers: [{ signer: referenceSigner, weight: 1 }],
-        thershold: 1,
-        metadata: {
-          name: 'Test'
-        }
-      })
-
-      // 2 signatures are made to publish signers
-      expect(referenceSigner.signingRequests).to.equal(2)
-
-      const signingRequestsBefore = referenceSigner.signingRequests
-
-      await initialAuthToFinish(session)
-
-      totalCount = 0
-
-      // Create a bunch of API clients sequentially
-      for (let i = 0; i < 10; i++) {
-        await expect(session.getAPI(networksWithApi[0])).to.be.rejected
-      }
-
-      expect(totalCount).to.equal(10)
-      expect(referenceSigner.signingRequests).to.equal(signingRequestsBefore + 1)
-    })
-
-    it('Should neither re-authenticate nor retry if request succeeds', async () => {
-      const referenceSigner = new CountingSigner(ethers.Wallet.createRandom())
-
-      const session = await Session.open({
-        context,
-        networks: networksWithApi,
-        referenceSigner: await referenceSigner.getAddress(),
-        signers: [{ signer: referenceSigner, weight: 1 }],
-        thershold: 1,
-        metadata: {
-          name: 'Test'
-        }
-      })
-
-      await initialAuthToFinish(session)
-
-      const api = await session.getAPI(networksWithApi[0])
-
-      const okResponses = [true]
-      server.post('/rpc/ArcadeumAPI/FriendList').thenCallback(async () => {
-        return { statusCode: okResponses.shift() ? 200 : 401, body: JSON.stringify({}) }
-      })
-
-      totalCount = 0
-
-      await expect(api.friendList({ page: {} })).to.be.fulfilled
-
-      // no re-authentication since it succeeded
-      expect(totalCount).to.equal(0)
+      expect(session.jwts[sequenceApiUrl]).to.be.undefined
     })
 
     describe('With expiration', () => {
@@ -1032,11 +873,11 @@ describe('Wallet integration', function () {
           }
         })
 
-        await initialAuthToFinish(session)
+        await Promise.all(session.authPromises.map((p) => p.promise))
 
         expect(totalCount).to.equal(1)
-        expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
-        expect(session._jwts.get(sequenceApiUrl)?.expiration).to.equal(baseTime + 240 - 60)
+        expect(session.jwts[sequenceApiUrl].expiration).to.equal(baseTime + 240 - 60)
+        expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
 
         // Force expire (1 hour)
         const newBaseTime = baseTime + (60 * 60)
@@ -1047,8 +888,8 @@ describe('Wallet integration', function () {
         await session.getAPI(networksWithApi[0])
 
         expect(totalCount).to.equal(2)
-        expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
-        expect(session._jwts.get(sequenceApiUrl)?.expiration).to.equal(newBaseTime + 240 - 60)
+        expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
+        expect(session.jwts[sequenceApiUrl].expiration).to.equal(newBaseTime + 240 - 60)
       })
 
       it("Should force min expiration time", async () => {
@@ -1069,11 +910,11 @@ describe('Wallet integration', function () {
           }
         })
 
-        await initialAuthToFinish(session)
+        await Promise.all(session.authPromises.map((p) => p.promise))
 
         expect(totalCount).to.equal(1)
-        expect(await getTokenForURL(session, sequenceApiUrl)).to.equal(fakeJwt)
-        expect(session._jwts.get(sequenceApiUrl)?.expiration).to.equal(baseTime + 120 - 60)
+        expect(session.jwts[sequenceApiUrl].expiration).to.equal(baseTime + 120 - 60)
+        expect(session.jwts[sequenceApiUrl].token).to.equal(fakeJwt)
       })
     })
   })